+++ packages/core/navigation-next/src/components/presentational/LayoutManager/ResizeControl.js
@@ -1,6 +1,12 @@
 // @flow
 
-import React, { PureComponent, Fragment, type Node, type Ref } from 'react';
+import React, {
+  PureComponent,
+  Fragment,
+  type Node,
+  type Ref,
+  type ElementRef,
+} from 'react';
 import raf from 'raf-schd';
 import {
   withAnalyticsEvents,
@@ -17,6 +23,7 @@ import {
   GLOBAL_NAV_WIDTH,
   CONTENT_NAV_WIDTH,
   CONTENT_NAV_WIDTH_COLLAPSED,
+  GLOBAL_NAV_COLLAPSE_THRESHOLD,
 } from '../../../common/constants';
 import { Shadow } from '../../../common/primitives';
 import PropertyToggle from './PropertyToggle';
@@ -27,6 +34,9 @@ const HANDLE_OFFSET = 4;
 const INNER_WIDTH = 20;
 const OUTER_WIDTH = INNER_WIDTH + HANDLE_OFFSET;
 const HANDLE_WIDTH = 2;
+const shouldResetGrabArea = (width: number) => {
+  return width >= GLOBAL_NAV_COLLAPSE_THRESHOLD && width < CONTENT_NAV_WIDTH;
+};
 
 const Outer = (props: *) => (
   <div css={{ position: 'relative', width: OUTER_WIDTH }} {...props} />
@@ -113,9 +123,9 @@ const Button = ({
   </button>
 );
 
-// tinker with the DOM directly by setting style properties, makes the
-function applyMutations(
-  elements: Array<{ property: string, ref: HTMLElement }>,
+// tinker with the DOM directly by setting style properties, updates the grab bar position by changing padding-left and width.
+function updateResizeAreaPosition(
+  elements: Array<{ property: 'padding-left' | 'width', ref: ElementRef<*> }>,
   width: number,
 ) {
   elements.forEach(({ property, ref }) => {
@@ -166,7 +176,10 @@ type Props = WithAnalyticsEventsProps & {
   flyoutIsOpen: boolean,
   isDisabled: boolean,
   mouseIsOverNavigation: boolean,
-  mutationRefs: Array<{ ref: HTMLElement, property: string }>,
+  mutationRefs: Array<{
+    ref: ElementRef<*>,
+    property: 'padding-left' | 'width',
+  }>,
   navigation: Object,
 };
 type State = {
@@ -233,7 +246,7 @@ class ResizeControl extends PureComponent<Props, State> {
 
   toggleCollapse = (trigger: string) => {
     const { navigation, createAnalyticsEvent } = this.props;
-    const newCollapsedState = !navigation.state.isCollapsed;
+    const newCollapsedState: boolean = !navigation.state.isCollapsed;
     navigation.toggleCollapse();
     navigationExpandedCollapsed(createAnalyticsEvent, {
       trigger,
@@ -309,7 +322,7 @@ class ResizeControl extends PureComponent<Props, State> {
     const width = initialWidth + delta;
 
     // apply updated styles to the applicable DOM nodes
-    applyMutations(mutationRefs, width);
+    updateResizeAreaPosition(mutationRefs, width);
 
     // NOTE: hijack the maual resize and force collapse, cancels mouse events
     if (event.screenX < window.screenX) {
@@ -322,27 +335,24 @@ class ResizeControl extends PureComponent<Props, State> {
   });
   handleResizeEnd = () => {
     const { navigation, createAnalyticsEvent } = this.props;
-    const { delta, didDragOpen, isDragging, width } = this.state;
-
-    let publishWidth = width;
-    let shouldCollapse;
+    const { delta, didDragOpen, isDragging, width: currentWidth } = this.state;
     const expandThreshold = 24;
-
     const resizerClicked = !isDragging && !this.invalidDragAttempted;
+    let publishWidth = currentWidth;
+    let shouldCollapse = false;
 
     // check if the intention was just a click, and toggle
     if (resizerClicked) {
-      publishWidth = Math.max(CONTENT_NAV_WIDTH, width);
+      publishWidth = Math.max(CONTENT_NAV_WIDTH, currentWidth);
       this.toggleCollapse('resizerClick');
     }
 
     // prevent the user from creating an unusable width
     if (publishWidth < CONTENT_NAV_WIDTH) {
       publishWidth = CONTENT_NAV_WIDTH;
-
       if (didDragOpen && delta > expandThreshold) {
         shouldCollapse = false;
-      } else {
+      } else if (currentWidth < GLOBAL_NAV_COLLAPSE_THRESHOLD) {
         shouldCollapse = true;
       }
     } else {
@@ -373,6 +383,10 @@ class ResizeControl extends PureComponent<Props, State> {
       productNavWidth: publishWidth,
       isCollapsed: shouldCollapse,
     });
+
+    if (shouldResetGrabArea(currentWidth)) {
+      updateResizeAreaPosition(this.props.mutationRefs, CONTENT_NAV_WIDTH);
+    }
 
     // cleanup
     window.removeEventListener('mousemove', this.handleResize);