+++ packages/core/navigation-next/src/components/presentational/LayoutManager/__tests__/unit/ResizeControl.js
@@ -2,7 +2,6 @@
 
 import React from 'react';
 import { mount, shallow } from 'enzyme';
-import cloneDeep from 'lodash.clonedeep';
 import ChevronLeft from '@atlaskit/icon/glyph/chevron-left';
 import { ResizeControlBase } from '../../ResizeControl';
 import { navigationExpandedCollapsed } from '../../../../../common/analytics';
@@ -12,45 +11,62 @@ jest.mock('../../../../../common/analytics', () => ({
 }));
 
 describe('ResizeControlBase', () => {
+  let props;
+
   beforeEach(() => {
     jest.clearAllMocks();
     jest.resetModules();
-  });
-
-  const resizeControlProps = {
-    collapseToggleTooltipContent: () => ({ text: '', char: '' }),
-    createAnalyticsEvent: (() => ({ fire: Function.prototype }): any),
-    expandCollapseAffordanceRef: { current: null },
-    experimental_flyoutOnHover: false,
-    flyoutIsOpen: false,
-    isDisabled: false,
-    mouseIsOverNavigation: false,
-    mutationRefs: [],
-    navigation: {
-      state: {
-        isCollapsed: false,
+    props = {
+      collapseToggleTooltipContent: () => ({ text: '', char: '' }),
+      createAnalyticsEvent: (() => ({ fire: Function.prototype }): any),
+      expandCollapseAffordanceRef: { current: null },
+      experimental_flyoutOnHover: false,
+      flyoutIsOpen: false,
+      isDisabled: false,
+      mouseIsOverNavigation: false,
+      mutationRefs: [
+        {
+          property: 'width',
+          ref: {
+            style: {
+              getPropertyValue: Function.prototype,
+              setProperty: Function.prototype,
+            },
+          },
+        },
+        {
+          property: 'padding-left',
+          ref: {
+            style: {
+              getPropertyValue: Function.prototype,
+              setProperty: Function.prototype,
+            },
+          },
+        },
+      ],
+      navigation: {
+        state: {
+          isCollapsed: false,
+          productNavWidth: 100,
+        },
+        manualResizeStart: jest.fn(),
+        manualResizeEnd: jest.fn(),
+        toggleCollapse: Function.prototype,
       },
-      manualResizeStart: Function.prototype,
-      toggleCollapse: Function.prototype,
-    },
-  };
+    };
+  });
 
   it('should render correctly', () => {
     const wrapper = shallow(
-      <ResizeControlBase {...resizeControlProps}>
-        {() => null}
-      </ResizeControlBase>,
+      <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
     );
 
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should call navigationExpandedCollapsed with chevronHover trigger when clicking on chevron while flyout is open', () => {
-    const props = cloneDeep(resizeControlProps);
-    props.navigation.state.isCollapsed = false;
-
     const wrapper = mount(
-      <ResizeControlBase {...resizeControlProps} flyoutIsOpen>
+      <ResizeControlBase {...props} flyoutIsOpen>
         {() => null}
       </ResizeControlBase>,
     );
@@ -59,7 +75,7 @@ describe('ResizeControlBase', () => {
 
     expect(navigationExpandedCollapsed).toHaveBeenCalledTimes(1);
     expect(navigationExpandedCollapsed).toHaveBeenCalledWith(
-      resizeControlProps.createAnalyticsEvent,
+      props.createAnalyticsEvent,
       {
         trigger: 'chevronHover',
         isCollapsed: true,
@@ -68,11 +84,9 @@ describe('ResizeControlBase', () => {
   });
 
   it('should call navigationExpandedCollapsed with chevron trigger when clicking on chevron while flyout is not open', () => {
-    const props = cloneDeep(resizeControlProps);
     props.navigation.state.isCollapsed = false;
-
     const wrapper = mount(
-      <ResizeControlBase {...resizeControlProps} flyoutIsOpen={false}>
+      <ResizeControlBase {...props} flyoutIsOpen={false}>
         {() => null}
       </ResizeControlBase>,
     );
@@ -81,12 +95,245 @@ describe('ResizeControlBase', () => {
 
     expect(navigationExpandedCollapsed).toHaveBeenCalledTimes(1);
     expect(navigationExpandedCollapsed).toHaveBeenCalledWith(
-      resizeControlProps.createAnalyticsEvent,
+      props.createAnalyticsEvent,
       {
         trigger: 'chevron',
         isCollapsed: true,
       },
     );
+  });
+
+  it('should call navigationExpandedCollapsed with resizerClick trigger when clicking recollapse on resize area', () => {
+    props.navigation.state.isCollapsed = false;
+    const wrapper = mount(
+      <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+    );
+
+    wrapper.instance().handleResizeEnd();
+
+    expect(navigationExpandedCollapsed).toHaveBeenCalledTimes(1);
+    expect(navigationExpandedCollapsed).toHaveBeenCalledWith(
+      props.createAnalyticsEvent,
+      {
+        trigger: 'resizerClick',
+        isCollapsed: true,
+      },
+    );
+  });
+
+  describe('when the component is resizing', () => {
+    describe('when starting to drag', () => {
+      it('should initialize dragging state', () => {
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState({ mouseIsDown: true, isDragging: false });
+
+        wrapper.instance().handleResize({ pageX: 100 });
+        requestAnimationFrame.step();
+
+        expect(wrapper.state('isDragging')).toEqual(true);
+        expect(wrapper.state('didDragOpen')).toEqual(false);
+        expect(wrapper.state('initialWidth')).toEqual(
+          props.navigation.state.productNavWidth,
+        );
+      });
+
+      it('should call navigation.manualResizeStart if isCollapsed is false', () => {
+        props.navigation.state.isCollapsed = false;
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState({ mouseIsDown: true, isDragging: false });
+
+        wrapper.instance().handleResize({ pageX: 100 });
+        requestAnimationFrame.step();
+
+        expect(props.navigation.manualResizeStart).toHaveBeenCalledTimes(1);
+        expect(props.navigation.manualResizeStart).toHaveBeenCalledWith(
+          props.navigation.state,
+        );
+      });
+
+      it('should call navigation.manualResizeStart if isCollapsed is true', () => {
+        props.navigation.state.isCollapsed = true;
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState({ mouseIsDown: true, isDragging: false });
+
+        wrapper.instance().handleResize({ pageX: 100 });
+        requestAnimationFrame.step();
+
+        expect(props.navigation.manualResizeStart).toHaveBeenCalledTimes(1);
+        expect(props.navigation.manualResizeStart).toHaveBeenCalledWith({
+          productNavWidth: 20,
+          isCollapsed: false,
+        });
+        expect(wrapper.state('didDragOpen')).toEqual(true);
+        expect(wrapper.state('initialWidth')).toEqual(20);
+      });
+    });
+
+    describe('when dragging', () => {
+      it('should not change width and delta when mouseIsDown is false', () => {
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState({ mouseIsDown: false, isDragging: true });
+        const { width: cachedWidth, delta: cachedDelta } = wrapper.state();
+
+        wrapper.instance().handleResize({ pageX: 100 });
+        requestAnimationFrame.step();
+
+        expect(wrapper.state('width')).toEqual(cachedWidth);
+        expect(wrapper.state('delta')).toEqual(cachedDelta);
+      });
+
+      it('should change width and delta when mouseIsDown is true', () => {
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState({
+          mouseIsDown: true,
+          isDragging: true,
+          initialWidth: 50,
+        });
+
+        wrapper.instance().handleResize({ pageX: 100 });
+        requestAnimationFrame.step();
+
+        expect(wrapper.state('width')).toEqual(150);
+        expect(wrapper.state('delta')).toEqual(100);
+      });
+
+      it('should change mutationRef style if new value is different than old value', () => {
+        const pageX = 100;
+        props.mutationRefs = [
+          {
+            property: 'padding-left',
+            ref: {
+              style: {
+                getPropertyValue: jest.fn().mockReturnValue('562px'),
+                setProperty: jest.fn(),
+              },
+            },
+          },
+          //not supposed to call setProperty for the ref below
+          {
+            property: 'width',
+            ref: {
+              style: {
+                getPropertyValue: jest.fn().mockReturnValue(`${pageX}px`),
+                setProperty: jest.fn(),
+              },
+            },
+          },
+        ];
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState({ mouseIsDown: true, isDragging: true });
+
+        wrapper.instance().handleResize({ pageX });
+        requestAnimationFrame.step();
+
+        expect(
+          props.mutationRefs[0].ref.style.getPropertyValue,
+        ).toHaveBeenCalledTimes(1);
+        expect(
+          props.mutationRefs[0].ref.style.setProperty,
+        ).toHaveBeenCalledWith('padding-left', '100px');
+        expect(
+          props.mutationRefs[1].ref.style.getPropertyValue,
+        ).toHaveBeenCalledTimes(1);
+        expect(
+          props.mutationRefs[1].ref.style.setProperty,
+        ).not.toHaveBeenCalled();
+      });
+    });
+
+    describe('when releasing drag', () => {
+      it('should collapse if dragged below collapse threshold', () => {
+        const state = { delta: -100, isDragging: true, width: 140 };
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState(state);
+
+        wrapper.instance().handleResizeEnd();
+
+        expect(props.navigation.manualResizeEnd).toHaveBeenCalledTimes(1);
+        expect(props.navigation.manualResizeEnd).toHaveBeenCalledWith({
+          productNavWidth: 240,
+          isCollapsed: true,
+        });
+      });
+
+      it('should resize back to default width if dragged above collapse threshold and below default width', () => {
+        const state = { delta: -15, isDragging: true, width: 225 };
+        props.mutationRefs = [
+          {
+            property: 'padding-left',
+            ref: {
+              style: {
+                getPropertyValue: jest.fn(),
+                setProperty: jest.fn(),
+              },
+            },
+          },
+          {
+            property: 'width',
+            ref: {
+              style: {
+                getPropertyValue: jest.fn(),
+                setProperty: jest.fn(),
+              },
+            },
+          },
+        ];
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState(state);
+
+        wrapper.instance().handleResizeEnd();
+
+        expect(props.navigation.manualResizeEnd).toHaveBeenCalledTimes(1);
+        expect(props.navigation.manualResizeEnd).toHaveBeenCalledWith({
+          productNavWidth: 240,
+          isCollapsed: false,
+        });
+        expect(
+          props.mutationRefs[0].ref.style.getPropertyValue,
+        ).toHaveBeenCalledTimes(1);
+        expect(
+          props.mutationRefs[0].ref.style.setProperty,
+        ).toHaveBeenCalledWith('padding-left', '240px');
+        expect(
+          props.mutationRefs[1].ref.style.getPropertyValue,
+        ).toHaveBeenCalledTimes(1);
+        expect(
+          props.mutationRefs[1].ref.style.setProperty,
+        ).toHaveBeenCalledWith('width', '240px');
+      });
+
+      it('should resize to greater width if dragged above default width', () => {
+        const state = { delta: 130, isDragging: true, width: 370 };
+        const wrapper = mount(
+          <ResizeControlBase {...props}>{() => null}</ResizeControlBase>,
+        );
+        wrapper.setState(state);
+
+        wrapper.instance().handleResizeEnd();
+
+        expect(props.navigation.manualResizeEnd).toHaveBeenCalledTimes(1);
+        expect(props.navigation.manualResizeEnd).toHaveBeenCalledWith({
+          productNavWidth: 370,
+          isCollapsed: false,
+        });
+      });
+    });
   });
 });