+++ packages/editor/editor-wikimarkup-transformer/src/__tests__/unit/parser/formatters/monospace.ts
@@ -37,6 +37,13 @@ another line`;
     expect(transformer.parse(wiki)).toMatchSnapshot();
   });
 
+  test('should detect monospace mark and ignoring the invalid closing symbol', () => {
+    const wiki = 'This is a {{monospace}}monospace}} text';
+
+    const transformer = new WikiMarkupTransformer();
+    expect(transformer.parse(wiki)).toMatchSnapshot();
+  });
+
   test('should not be a monospace mark if there is a space after opening {{', () => {
     const wiki = 'This is not a {{ monospace}} text';