+++ packages/people-and-teams/focused-task-close-account/src/components/DeleteUserOverviewScreen/types.ts
@@ -1,0 +1,20 @@
+import { User } from '../../types';
+
+export interface DeleteUserOverviewScreenProps {
+  accessibleSites: AvailableSitesResponse;
+  isCurrentUser: boolean;
+  user: User;
+}
+
+interface AvailableSitesResponse {
+  sites: Sites[];
+}
+
+interface Sites {
+  cloudId: string;
+  url: string;
+  products: string[];
+  avatarUrl: string;
+  displayName: string;
+  isVertigo: boolean;
+}